Updates Site 53's first and second Z-levels
Adds additions to the D-Class Funzone, More Cameras and a window for Guards to observe through with shutters.

Moves SCP-529's Containment cell to the top of the Archive since its a better pet then the Parrot and since its safe. (Can Free roam the site a bit)

Replaces SCP-529's old containment room with a General Purpose Containment room that is empty.