Fix Fax Machines, Fixes issue #507
Hey this is the IT department, sorry about the fax machines.. Everything's been on back-order, but they've just arrived today and the new updates were ran out by Jared.
Turns out SCP-008's cremator was installed backwards, this wasn't a IT issue, but we went ahead and fixed it as well.
Thanks for your patience, - IT Department