Cleans up Train design, animates tunnel movement, adds death triggers, and comments out unused escape pod code.
Changed the train design up to include three middle carts, two junction platforms, a caboose, and a command car, amongst other internal QOL additions.

Animated the train tunnel to include moving plating, flooring, and walls. Does not use Plane_masters as those are beyond me and I think have the functionality and semi-good appearance of the Train is more important than a super high detail tunnel.

Removes the Scout Prep Area to make room for an extended Train Station. If this is a big issue I can map it back in, but as I saw it was mostly unused.

Commented out unused escape pod code. We don't have any escape pods anymore, this is fully implemented as a Shuttle.